+++ lib/montage/commands/generate.rb
@@ -17,7 +17,19 @@ module Montage
       #   The arguments given on the command line.
       #
       def self.run(argv)
-        new(Montage::Project.find(Dir.pwd),
+        # If there are any arguments, the first one is a path to a montage
+        # config file.
+        if argv.first and not Pathname.new(argv.first).file?
+          say color(<<-ERROR.compress_lines, :red)
+            Couldn't find `#{argv.first}' configuration file. Are you
+            sure you got the path right?
+          ERROR
+
+          exit(1)
+        end
+
+
+        new(Montage::Project.find(argv.first || Dir.pwd),
             Montage::Commands.config[:force]).run!
 
       rescue Montage::MissingProject